Mark some tests as slow
