+++ py-polars/tests/unit/io/test_iceberg.py
@@ -25,6 +25,8 @@ def iceberg_path(io_files_path: Path) -> str:
     return f"file://{iceberg_path.resolve()}"
 
 
+@pytest.mark.slow()
+@pytest.mark.write_disk()
 @pytest.mark.filterwarnings(
     "ignore:No preferred file implementation for scheme*:UserWarning"
 )
@@ -38,6 +40,8 @@ def test_scan_iceberg_plain(iceberg_path: str) -> None:
     }
 
 
+@pytest.mark.slow()
+@pytest.mark.write_disk()
 @pytest.mark.filterwarnings(
     "ignore:No preferred file implementation for scheme*:UserWarning"
 )
@@ -71,6 +75,8 @@ def test_scan_iceberg_filter_on_partition(iceberg_path: str) -> None:
         assert res.collect().rows() == [(1,), (3,)]
 
 
+@pytest.mark.slow()
+@pytest.mark.write_disk()
 @pytest.mark.filterwarnings(
     "ignore:No preferred file implementation for scheme*:UserWarning"
 )
