+++ py-polars/tests/unit/utils/test_show_versions.py
@@ -1,8 +1,11 @@
 from typing import Any
 
+import pytest
+
 import polars as pl
 
 
+@pytest.mark.slow()
 def test_show_versions(capsys: Any) -> None:
     pl.show_versions()
